+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v2.1.1
 ## 05/17/2020
 
+1. [](#new)
+    * Added `No User` option to allow disabling the username requirement. This is useful for when you have a token and the user is not required. (#166, thanks GwynethLlewelyn) 
 1. [](#bugfix)
     * Fixed classes not being loaded in `cli` commands due to Grav changes (#167)
     * Updated dependencies / recompiled JS for production