+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v2.1.2
+## 07/17/2020
+
+1. [](#bugfix)
+    * Fixed regression wrongly returning the installed Git version and causing all sort of problems, including unrelated histories not kicking off (#61, #168, #171, #173)
 # v2.1.1
 ## 05/17/2020